// packages/core/src/types/ethnicity.ts
export type Ethnicity =
  | 'west_africa'
  | 'east_africa'
  | 'north_africa'
  | 'central_africa'
  | 'southern_africa'
  | 'western_europe'
  | 'eastern_europe'
  | 'middle_east'
  | 'south_asia'
  | 'east_asia'
  | 'latin_america'
  | 'mixed';

export const ETHNICITIES: Ethnicity[] = [
  'west_africa',
  'east_africa',
  'north_africa',
  'central_africa',
  'southern_africa',
  'western_europe',
  'eastern_europe',
  'middle_east',
  'south_asia',
  'east_asia',
  'latin_america',
  'mixed',
];